Refactor: pet 경로 수정
